move to `for-pages` dir
