+++ packages/web/src/components/for-pages/liquidity/Statistics/index.spec.tsx
