+++ packages/web/src/components/for-pages/liquidity/TokenSymbol/index.spec.tsx
