+++ packages/web/src/components/for-pages/liquidity/TokenSymbol/index.tsx
