+++ packages/web/src/components/for-pages/liquidity/Withdraw/index.tsx
@@ -3,8 +3,8 @@ import React, { ChangeEvent, useCallback } from 'react'
 import { useState } from 'react'
 import { toAmountNumber, toNaturalNumber } from 'src/fixtures/utility'
 import styled from 'styled-components'
-import { totalStakedFor, unstakeQuery } from '../../fixtures/geyser/client'
-import { useUnstake } from '../../fixtures/geyser/hooks'
+import { totalStakedFor, unstakeQuery } from '../../../../fixtures/for-pages/liquidity/geyser/client'
+import { useUnstake } from '../../../../fixtures/for-pages/liquidity/geyser/hooks'
 import { Gap } from '../Gap'
 import { LargeInput } from '../LargeInput'
 import { Max } from '../Max'
