+++ packages/web/src/fixtures/for-pages/liquidity/geyser/__mocks__/hooks.ts
