+++ packages/web/src/fixtures/for-pages/liquidity/geyser/abi.ts
