+++ packages/web/src/fixtures/for-pages/liquidity/uniswap-pool/__mocks__/hooks.ts
