+++ packages/web/src/fixtures/for-pages/liquidity/uniswap-pool/abi.ts
