+++ packages/web/src/fixtures/for-pages/liquidity/uniswap-pool/cache-path.ts
