+++ packages/web/src/fixtures/for-pages/liquidity/uniswap-pool/hooks.ts
